elevadores/configuracoes/requirements.txt
bmp280==0.0.4
smbus2==0.4.1
pyserial==3.5